fix getPerformanceData

+++ user.h
@@ -20,7 +20,7 @@ int mkdir(const char*);
 int chdir(const char*);
 int dup(int);
 int getpid(void);
-int getPerformanceData(int * , int *);
+int getPerformanceData(long long int * , long long int *);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
